requirements.txt
fastapi==0.95.1
PyJWT==2.6.0
PyMySQL==1.0.3
uvicorn==0.21.1
replicate==0.7.0
Jinja2==3.1.0
python-multipart==0.0.6
boto3==1.26.133
email-validator==2.0.0.post2
